Orginiation of code, new sprites to be added into game, and default settings.cfg because i need to add code to autocreate it.
new sprites to be added into game, and default settings.cfg because i need to add code to autocreate it.
