+++ assets/maps/bit_tileset_test.tsx
