+++ src/bit_loadfile.h
@@ -1,0 +1,3 @@
+// Persistent storage functions
+static bool SaveStorageValue(unsigned int position, int value);
+static int LoadStorageValue(unsigned int position);
