+++ src/bit_tmxload.c
